比对,检测，word公式解析

+++ spring-support-common-starter/src/main/java/com/chua/starter/common/support/logger/SysLog.java
@@ -58,7 +58,7 @@ public class SysLog {
     /**
      * 耗时(s)
      */
-    private Long logCost;
+    private Double logCost;
     /**
      * 請求地址
      */
